هندل کردن تخفیف در باشگاه مشتریان + اصلاح درصد تخفیف

+++ Prototype/OODProjectReact/Models/Good.cs
@@ -16,7 +16,7 @@ namespace OODProjectReact.Models
         public string Name { get; set; }
         public int Price { get; set; }
         public int Quantity { get; set; }
-        public int Discount { get; set; }
+        public byte Discount { get; set; }
         public string Explanation { get; set; }
 
         public virtual ICollection<GoodCategory> GoodCategory { get; set; }
